backend/requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
Django==5.0.6
django-redis==5.4.0
django-comment-migrate==0.1.7
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
loguru==0.7.2
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
paho-mqtt==2.1.0
platformdirs==4.2.1
psycopg2==2.9.10
PyJWT==2.8.0
setuptools
sqlparse==0.5.0
django-timescaledb==0.2.13
urllib3==2.2.1
pika==1.3.2
minio==7.2.0


